+++ src/mol-util/input/input-observer.ts
@@ -3,6 +3,7 @@
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author David Sehnal <david.sehnal@gmail.com>
+ * @author Russell Parker <russell@benchling.com>
  */
 
 import { Subject, Observable } from 'rxjs';
@@ -360,6 +361,7 @@ namespace InputObserver {
         let isInside = false;
         let hasMoved = false;
 
+        const resizeObserver: ResizeObserver = new window.ResizeObserver(onResize);
         const events = createEvents();
         const { drag, interactionEnd, wheel, pinch, gesture, click, move, leave, enter, resize, modifiers, key, keyUp, keyDown, lock } = events;
 
@@ -393,7 +395,7 @@ namespace InputObserver {
             document.addEventListener('pointerlockchange', onPointerLockChange, false);
             document.addEventListener('pointerlockerror', onPointerLockError, false);
 
-            window.addEventListener('resize', onResize, false);
+            resizeObserver.observe(element.parentElement!);
         }
 
         function dispose() {
@@ -423,9 +425,10 @@ namespace InputObserver {
             document.removeEventListener('pointerlockchange', onPointerLockChange, false);
             document.removeEventListener('pointerlockerror', onPointerLockError, false);
 
-            window.removeEventListener('resize', onResize, false);
-
             cross.remove();
+
+            resizeObserver.unobserve(element.parentElement!);
+            resizeObserver.disconnect();
         }
 
         function onPointerLockChange() {
@@ -773,7 +776,17 @@ namespace InputObserver {
             gestureDelta(ev, true);
         }
 
-        function onResize(ev: Event) {
+        function onMouseEnter(ev: Event) {
+            isInside = true;
+            enter.next(void 0);
+        }
+
+        function onMouseLeave(ev: Event) {
+            isInside = false;
+            leave.next(void 0);
+        }
+
+        function onResize() {
             resize.next({});
         }